+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `nativeBigInt256` module
   - `String.padStart()` requires TS target ES2017 or higher
   - `BigInt` requires TS target ES2020 or higher
+- `index.ts` with `LexOrder` factory
 
 
 [unreleased]: https://github.com/aryelgois/lexorder/compare/initial-commit...HEAD